Move server.loggingPlatform to its own file
